// src/app/summoner/models/champion-stats.ts
export class ChampionStats {
  totalOfMinionsKilled: number;
  totalMinionsKilled: number;
  neutralMinionsKilled: number;
  neutralMinionsKilledEnemyJungle: number;
  neutralMinionsKilledTeamJungle: number;
  minionsKilledPerMin: number;
  gameDurationSeconds: number;
  gameDurationMinutes: number;
  golds: number;
  kills: number;
  deaths: number;
  assists: number;
  kda: number;
  numberOfGames: number;
}